requirements.txt
opencv-python
pytesseract
matplotlib
spacy
pandas
geopandas
geopy
google-cloud-vision

